[OBEY-CAMPAIGN-8deed8b4] refactor: modularize quest metadata updates
